pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "proofgraph"
version = "0.1.0"
description = "Agent-built, source-cited evidence graphs for security control artifacts."
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12",
  "pydantic>=2.6",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
proofgraph = "proofgraph.cli:app"

[tool.setuptools.packages.find]
include = ["proofgraph*"]
exclude = ["profiles*", "tests*", "examples*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
